Session based Login
Changed UserController to use session based system. This was written by Ammar. Issue with pull request, we chose to copy the code directly, this is functioning as it did on the development branch.
Removed review page from navabar, making it only accessable via the Leave a Review button on dynamic product pages - made said button only viewable/interactable if users are logged in. Added error message to review page if user is not logged in. Review form now detects if user is logged in and uses their user_id if they are