Previous Orders
previous orders page UI changed and made dynamic based on logged in user as well as what the user has in order table (uses foreign keys in table to get data from other tables also). If no user is logged in, trying to access page redirects to login page with error message. Gave the ReOrder button functionality. Changed Order model to allow total_price column to be fillable, respective seeder data was changed to have total_price values that are accurate and price of items was made accurate as well.