Basket update
Added update basket button that is attached to each cart-item (which are now forms). Button takes basket_id and quantity and saves the input quantity on the page to the cart_table.
Gives user a message when they update.
Made adding to cart send you to the login screen with appropriate message, if not logged in.
If trying to access /delete/{id} URL then sent to login page, apt message.
If trying to access /basket when not logged in sends to login page, apt message.