Home rework step 2
Tried to change the select to a list of radio buttons, however trying to style them was difficult so decided to stay with select
Gave select class="category-select", changing option's hover to gold was unsuccessful so it has stayed as browser/OS default
Gave the placeholder option of select disabled, selected and hidden as attributes, this selects it but hides in option when the user will select a category
Added h2 to productFilters section to create more noticeable change to page, gave it the styling of h2 in previous section
Gave search button the same styling as the hero btn
Changed section class="products" to section class="products-carousel" for better differentiation

Gave the select on product page the same class as the one on home for styling reasons. Did the same changes to its placeholder

Noticed master page was no longer displaying the dynamic basket count. Fixed issue